requirements.txt
accelerate==1.1.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
appnope==0.1.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
comm==0.2.2
datasets==3.0.1
debugpy==1.8.6
decorator==5.1.1
dill==0.3.8
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.21.0
filelock==3.16.1
frozenlist==1.4.1
fsspec==2024.6.1
graphviz==0.20.3
huggingface-hub==0.25.1
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.17
nbformat==5.10.4
nbstripout==0.8.1
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rpds-py==0.21.0
safetensors==0.4.5
six==1.16.0
stack-data==0.6.3
sympy==1.13.3
tokenizers==0.20.0
torch==2.2.2
torchviz==0.0.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.45.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.13.1
zipp==3.20.2
